Move types into related generator spaces.
