+++ src/manifest/resource/Repository/types.ts
@@ -1,0 +1,14 @@
+import { ConfigCollection, DefinitionObject } from '../../types';
+
+/**
+ * Base information defining a repository for use.
+ */
+export interface RepositoryDefinition extends DefinitionObject {
+  readonly id: string;
+  readonly url: string;
+}
+
+/**
+ * A collection of Repository configuration definitions.
+ */
+export type RepositoryCollection = ConfigCollection<RepositoryDefinition>;
